+++ previous-versions/assess/pkg/models/shared/bankingtransactionref.go
@@ -3,14 +3,15 @@
 package shared
 
 import (
-	"github.com/codatio/client-sdk-go/previous-versions/assess/pkg/types"
+	"github.com/codatio/client-sdk-go/previous-versions/assess/pkg/utils"
+	"github.com/ericlagergren/decimal"
 )
 
 type BankingTransactionRef struct {
-	AccountID        *string        `json:"accountId,omitempty"`
-	AccountName      *string        `json:"accountName,omitempty"`
-	Amount           *types.Decimal `json:"amount,omitempty"`
-	DataConnectionID *string        `json:"dataConnectionId,omitempty"`
+	AccountID        *string      `json:"accountId,omitempty"`
+	AccountName      *string      `json:"accountName,omitempty"`
+	Amount           *decimal.Big `decimal:"number" json:"amount,omitempty"`
+	DataConnectionID *string      `json:"dataConnectionId,omitempty"`
 	// In Codat's data model, dates and times are represented using the <a class="external" href="https://en.wikipedia.org/wiki/ISO_8601" target="_blank">ISO 8601 standard</a>. Date and time fields are formatted as strings; for example:
 	//
 	// ```
@@ -35,6 +36,17 @@ type BankingTransactionRef struct {
 	ID          *string `json:"id,omitempty"`
 }
 
+func (b BankingTransactionRef) MarshalJSON() ([]byte, error) {
+	return utils.MarshalJSON(b, "", false)
+}
+
+func (b *BankingTransactionRef) UnmarshalJSON(data []byte) error {
+	if err := utils.UnmarshalJSON(data, &b, "", false, false); err != nil {
+		return err
+	}
+	return nil
+}
+
 func (o *BankingTransactionRef) GetAccountID() *string {
 	if o == nil {
 		return nil
@@ -49,7 +61,7 @@ func (o *BankingTransactionRef) GetAccountName() *string {
 	return o.AccountName
 }
 
-func (o *BankingTransactionRef) GetAmount() *types.Decimal {
+func (o *BankingTransactionRef) GetAmount() *decimal.Big {
 	if o == nil {
 		return nil
 	}