+++ previous-versions/assess/pkg/models/shared/transactioncategory.go
@@ -3,17 +3,29 @@
 package shared
 
 import (
-	"github.com/codatio/client-sdk-go/previous-versions/assess/pkg/types"
+	"github.com/codatio/client-sdk-go/previous-versions/assess/pkg/utils"
+	"github.com/ericlagergren/decimal"
 )
 
 type TransactionCategory struct {
 	// Returns the confidence of the suggested category for the transaction. The value is between 0 and 100.
-	Confidence *types.Decimal `json:"confidence,omitempty"`
+	Confidence *decimal.Big `decimal:"number" json:"confidence,omitempty"`
 	// The suggested category is an ordered array of category levels where each element (or level) is a subcategory of the previous element (or level).
 	Levels []string `json:"levels,omitempty"`
 }
 
-func (o *TransactionCategory) GetConfidence() *types.Decimal {
+func (t TransactionCategory) MarshalJSON() ([]byte, error) {
+	return utils.MarshalJSON(t, "", false)
+}
+
+func (t *TransactionCategory) UnmarshalJSON(data []byte) error {
+	if err := utils.UnmarshalJSON(data, &t, "", false, false); err != nil {
+		return err
+	}
+	return nil
+}
+
+func (o *TransactionCategory) GetConfidence() *decimal.Big {
 	if o == nil {
 		return nil
 	}