+++ previous-versions/assess/RELEASES.md
@@ -381,3 +381,13 @@ Based on:
 - [go v0.25.1] previous-versions/assess
 ### Releases
 - [Go v0.25.1] https://github.com/codatio/client-sdk-go/releases/tag/previous-versions/assess/v0.25.1 - previous-versions/assess
+
+## 2023-09-20 00:12:43
+### Changes
+Based on:
+- OpenAPI Doc 1.0 https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Assess.yaml
+- Speakeasy CLI 1.88.0 (2.118.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [go v0.25.2] previous-versions/assess
+### Releases
+- [Go v0.25.2] https://github.com/codatio/client-sdk-go/releases/tag/previous-versions/assess/v0.25.2 - previous-versions/assess